backend/pyproject.toml
[project]
name = "palleto-backend"
version = "0.1.0"
description = "FastAPI backend for the Palleto mobile MVP."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "firebase-admin>=6.5.0,<7.0.0",
  "python-multipart>=0.0.12,<1.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.0.0,<9.0.0",
  "ruff>=0.7.0,<1.0.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
